# server/static/openapi.yaml
openapi: 3.0.3
info:
  title: AgriLink API
  description: >
    REST API for AgriLink — a platform connecting farmers with agricultural
    experts through posts, topic-based communities, and direct messaging.
    Authentication uses JWT bearer tokens obtained from /api/auth/login.
  version: 1.0.0
servers:
  - url: https://agrilink-11rw.onrender.com
    description: Production (Render)
  - url: http://localhost:5000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/users/search:
    get:
      tags: [users]
      summary: Search users by username
      parameters: [{ name: q, in: query, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /api/users/{user_id}:
    get:
      tags: [users]
      summary: Get user profile
      parameters: [{ name: user_id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: OK }, "404": { description: Not found } }
    patch:
      tags: [users]
      summary: Update own profile
      parameters: [{ name: user_id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: OK }, "403": { description: Forbidden } }
    delete:
      tags: [users]
      summary: Delete account (self or admin)
      parameters: [{ name: user_id, in: path, required: true, schema: { type: integer } }]
      responses: { "204": { description: Deleted } }
  /api/users/{user_id}/follow:
    post:
      tags: [users]
      summary: Follow a user
      parameters: [{ name: user_id, in: path, required: true, schema: { type: integer } }]
      responses: { "201": { description: Followed } }
    delete:
      tags: [users]
      summary: Unfollow a user
      parameters: [{ name: user_id, in: path, required: true, schema: { type: integer } }]
      responses: { "204": { description: Unfollowed } }
  /api/users/{user_id}/followers:
    get:
      tags: [users]
      summary: List a user's followers
      parameters: [{ name: user_id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: OK } }
  /api/users/{user_id}/following:
    get:
      tags: [users]
      summary: List accounts a user follows
      parameters: [{ name: user_id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: OK } }

  # ------------------------------------------------------------------ posts
...
  /api/communities/{community_id}:
    delete:
      tags: [communities]
      summary: Delete a community (owner/admin)
      parameters: [{ name: community_id, in: path, required: true, schema: { type: integer } }]
      responses: { "204": { description: Deleted } }
  /api/communities/{community_id}/join:
    post:
      tags: [communities]
      summary: Join a community
      parameters: [{ name: community_id, in: path, required: true, schema: { type: integer } }]
      responses: { "201": { description: Joined } }
  /api/communities/{community_id}/leave:
    delete:
      tags: [communities]
      summary: Leave a community
      parameters: [{ name: community_id, in: path, required: true, schema: { type: integer } }]
      responses: { "204": { description: Left } }
  /api/communities/{community_id}/members:
    get:
      tags: [communities]
      summary: List community members
      parameters: [{ name: community_id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: OK } }
  /api/communities/{community_id}/posts:
    get:
      tags: [communities]
      summary: List posts in a community
      parameters: [{ name: community_id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: OK } }

  # ---------------------------------------------------------------- messages
...
  /api/messages/user/{user_id}:
    get:
      tags: [messages]
      summary: Message history with a user
      parameters: [{ name: user_id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: OK } }
...